remove previous runnable files and keep the refacted files
1. all the test cases are runnable
