--- udao_spark/optimizer/refactored/moo_algos/__init__.py
@@ -1,11 +1,0 @@
-# Copyright (c) 2020 École Polytechnique
-#
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-#
-# Author(s): Qi FAN <qi dot fan at polytechnique dot edu>
-#
-# Description: TODO
-#
-# Created at 18/04/2024
